+++ astropy/coordinates/representation.py
@@ -290,6 +290,60 @@ class CartesianRepresentation(BaseRepresentation):
     def to_cartesian(self):
         return self
 
+    def transform(self, matrix):
+        """
+        Transform the cartesian coordinates using a 3x3 matrix.
+
+        This returns a new representation and does not modify the original one.
+
+        Parameters
+        ----------
+        matrix : `~numpy.ndarray`
+            A 3x3 transformation matrix, such as a rotation matrix.
+
+        Examples
+        --------
+
+        We can start off by creating a cartesian representation object:
+
+            >>> from astropy import units as u
+            >>> from astropy.coordinates import CartesianRepresentation
+            >>> rep = CartesianRepresentation([1, 2] * u.pc,
+            ...                               [2, 3] * u.pc,
+            ...                               [3, 4] * u.pc)
+
+        We now create a rotation matrix around the z axis:
+
+            >>> from astropy.coordinates.angles import rotation_matrix
+            >>> rotation = rotation_matrix(30 * u.deg, axis='z')
+
+        Finally, we can apply this transformation:
+
+            >>> rep_new = rep.transform(rotation)
+            >>> rep_new.xyz  # doctest: +FLOAT_CMP
+            <Quantity [[ 1.8660254 , 3.23205081],
+                       [ 1.23205081, 1.59807621],
+                       [ 3.        , 4.        ]] pc>
+        """
+
+        # TODO: since this is likely to be a widely used function in coordinate
+        # transforms, it should be optimized (for example in Cython).
+
+        # Get xyz once since it's an expensive operation
+        xyz = self.xyz
+
+        # Since the underlying data can be n-dimensional, reshape to a
+        # 2-dimensional (3, N) array.
+        vec = xyz.reshape((3, xyz.size // 3))
+
+        # Do the transformation
+        vec_new = np.dot(np.asarray(matrix), vec)
+
+        # Restore the original shape
+        vec_new = vec_new.reshape(xyz.shape)
+
+        return self.__class__(*vec_new)
+
 
 class UnitSphericalRepresentation(BaseRepresentation):
     """